# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Quran Semantic Research API",
    "description": "واجهة تأسيسية لمنصة الاستقراء الدلالي.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/ayahs/{surah_number}/{ayah_number}": {
      "get": {
        "tags": [
          "Quran"
        ],
        "summary": "Get Ayah",
        "operationId": "get_ayah_api_v1_ayahs__surah_number___ayah_number__get",
        "parameters": [
          {
            "name": "surah_number",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 114,
              "minimum": 1,
              "title": "Surah Number"
            }
          },
          {
            "name": "ayah_number",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 286,
              "minimum": 1,
              "title": "Ayah Number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}